LayoutTests/compositing/geometry/object-clip-rects-assertion.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN"
  "http://www.w3.org/TR/html4/strict.dtd">
<html>
<head>
  <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
  <title>Clip Rects assertion</title>
  <style type="text/css" media="screen">

    .container {
        position: absolute;
        overflow: hidden;
        padding: 10px;
        border: 1px solid black;
    }
    
    .box {
      width: 100px;
      height: 100px;
      background-color: blue;
      opacity: 0.8;
    }
    
  </style>
  <script type="text/javascript" charset="utf-8">
    if (window.testRunner)
      testRunner.dumpAsText();
  </script>
</head>
<body>

    <p>This testcase should not fire an assertion related to clip rects.</p>
    <div class="container">
        <object height="100" width="100">
            <param value="../resources/video.mp4" name="src">
            <param value="false" name="autoplay">
        </object>
        <div class="box">
        </div>
    </div>
    
</body>
</html>
